provide mechanisms to forward output from jnlp app
A misbehaving application may log information to stderr, or even stdout.
Sometimes it is useful to collect this information for diagnostic
purposes. Provide a mechanism to capture this output in a customizable
fashion using callback functions invoked on a goroutine with the
stdin/stderr pipe returned from the exec.Cmd we launch.